Fix PY_SSIZE_T_CLEAN compatibility in C extension modules
This commit fixes segmentation faults and ensures compatibility with
Python 3.8-3.12+ by properly handling the PY_SSIZE_T_CLEAN macro.

When PY_SSIZE_T_CLEAN is defined, the s# format specifier in
PyArg_ParseTuple requires Py_ssize_t* for the length parameter,
not int*. On 64-bit systems, Py_ssize_t is 8 bytes while int is
4 bytes, causing stack corruption when int* is incorrectly used.

Changes:
- ecmodule.c: Fix msg_len, pf_len, ps_len types for s# format
- pairingmodule.c: Fix b_len type (size_t -> Py_ssize_t)
- pairingmodule2.c (miracl): Fix b_len, m_len, c_len types
- pairingmodule3.c (relic): Fix string_len type
- stream_template.c: Add PY_SSIZE_T_CLEAN, fix keylen/len types
- XOR.c: Add PY_SSIZE_T_CLEAN before Python.h include
- Various headers: Ensure PY_SSIZE_T_CLEAN defined before Python.h
- Remove Python 2.x compatibility code from headers
- Update deprecated OpenSSL APIs to use EVP interface

This fixes the segfaults in pkenc_test.py and pksig_test.py that
were occurring in EC_CS98Test and ECDSATest.

+++ charm/core/math/pairing/miracl/pairingmodule2.c
@@ -435,7 +435,7 @@ int Element_init(Element *self, PyObject *args, PyObject *kwds)
 int Pairing_init(Pairing *self, PyObject *args, PyObject *kwds)
 {
 	char *params = NULL, *param_string = NULL;
-	size_t b_len = 0;
+	Py_ssize_t b_len = 0;
 	int aes_sec = -1;
     static char *kwlist[] = {"aes_sec", "params", "param_string", NULL};
 	
@@ -1568,7 +1568,7 @@ PyObject *AES_Encrypt(Element *self, PyObject *args)
 {
 	PyObject *keyObj = NULL, *tmp_obj = NULL; // string or bytes object
 	char *messageStr;
-	int m_len = 0;
+	Py_ssize_t m_len = 0;
 
 	if(!PyArg_ParseTuple(args, "Os#", &keyObj, &messageStr, &m_len)) {
 		PyErr_SetString(ElementError, "invalid arguments.");
@@ -1588,7 +1588,7 @@ PyObject *AES_Encrypt(Element *self, PyObject *args)
 		// perform AES encryption using miracl
 		char *cipher = NULL;
 
-		int c_len = aes_encrypt(keyStr, messageStr, m_len, &cipher);
+		int c_len = aes_encrypt(keyStr, messageStr, (int) m_len, &cipher);
 
 		PyObject *str = PyBytes_FromStringAndSize((const char *) cipher, c_len);
 		free(cipher);
@@ -1604,7 +1604,7 @@ PyObject *AES_Decrypt(Element *self, PyObject *args)
 {
 	PyObject *keyObj = NULL, *tmp_obj = NULL; // string or bytes object
 	char *ciphertextStr;
-	int c_len = 0;
+	Py_ssize_t c_len = 0;
 
 	if(!PyArg_ParseTuple(args, "Os#", &keyObj, &ciphertextStr, &c_len)) {
 		PyErr_SetString(ElementError, "invalid arguments.");
@@ -1619,7 +1619,7 @@ PyObject *AES_Decrypt(Element *self, PyObject *args)
 		// perform AES encryption using miracl
 		char *message = NULL;
 
-		int m_len = aes_decrypt(keyStr, ciphertextStr, c_len, &message);
+		int m_len = aes_decrypt(keyStr, ciphertextStr, (int) c_len, &message);
 
 		PyObject *str = PyBytes_FromStringAndSize((const char *) message, m_len);
 		free(message);
